requirements.txt
google-genai==1.17.0
dateparser
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
faiss-cpu
sentence-transformers
flask
python-dotenv